requirements/base.txt
celery==5.2.7
importlib-metadata==4.13.0
django-cors-headers==4.0.0
django-filter==23.2
djangorestframework==3.14.0
entrypoints==0.4
Pillow==9.5.0
pip==23.1.2
psycopg2-binary==2.9.6
python-dotenv==0.21.1